Move stream package to top level
