+++ stream/iterator.go
