+++ stream/operator.go
