+++ stream/project_test.go
@@ -6,9 +6,9 @@ import (
 	"testing"
 
 	"github.com/genjidb/genji/document"
-	"github.com/genjidb/genji/document/stream"
 	"github.com/genjidb/genji/sql/parser"
 	"github.com/genjidb/genji/sql/query/expr"
+	"github.com/genjidb/genji/stream"
 	"github.com/stretchr/testify/require"
 )
