+++ stream/stream.go
